Refactor achievements to use user ids instead of user objects
