+++ back-end/src/util/achievements/achievements.ts
@@ -1,7 +1,7 @@
 import { Achievement } from "./abstract-classes";
-import { User } from "../../db/entity";
-import { ApiError } from "../errorHandling/apiError";
-import { HttpResponseCode } from "../errorHandling/httpResponseCode";
+import { User } from "../../db/entity/hub";
+import { ApiError } from "../errorHandling";
+import { HttpResponseCode } from "../errorHandling";
 import { BeekeeperAchievement } from "./models";
 
 /**
@@ -20,6 +20,15 @@ export abstract class Achievements {
    */
   public static getAchievements(): Achievement[] {
     return this.achievementsCollection;
+  }
+
+  /**
+   * Returns achievement with given id
+   */
+  public static getAchievementWithId(id: string): Achievement {
+    return this.achievementsCollection.find((achievement: Achievement) =>
+      achievement.id === id
+    );
   }
 
   /**
@@ -52,6 +61,12 @@ export abstract class Achievements {
     // find in achievementsCollection
     // call incrementProgress
     // update data in cache
+    throw new ApiError(HttpResponseCode.NOT_IMPLEMENTED);
+  }
+
+  public static refreshUserProgressForAchievement(user: User, achievementId: string) {
+    // find in cache
+    // call sync()
     throw new ApiError(HttpResponseCode.NOT_IMPLEMENTED);
   }
 }
