+++ back-end/src/util/achievements/models/beekeeperAchievement.ts
@@ -1,10 +1,11 @@
 import { Achievement } from "../abstract-classes";
-import { ApiError } from "../../errorHandling/apiError";
-import { HttpResponseCode } from "../../errorHandling/httpResponseCode";
-import { User } from "../../../db/entity";
+import { ApiError } from "../../errorHandling";
+import { HttpResponseCode } from "../../errorHandling";
+import { User } from "../../../db/entity/hub";
 
 /**
  * The "Beekeeper" achievement
+ * Can be completed by finding all bees in the treasure hunt
  */
 export class BeekeeperAchievement extends Achievement {
   public id: string;
@@ -13,10 +14,10 @@ export class BeekeeperAchievement extends Achievement {
   public prizes: string;
   public finishMessage: string;
   public maxProgress: string;
-  public incrementProgress(user: User, token: string): number {
+  public incrementProgress(userId: number, token: string): number {
     throw new ApiError(HttpResponseCode.NOT_IMPLEMENTED);
   }
-  public checkProgress(user: User): number {
+  public checkProgress(userId: number): number {
     throw new ApiError(HttpResponseCode.NOT_IMPLEMENTED);
   }
 }
